Make sure buffer is writeable before writing, then save changes
Ideally, the buffer we're using to store our username-password
information is read-only, so that users don't make accidental changes,
deletions etc. to other parts of the document, which likely contain
sensitive information (mainly other passwords.)

Hence, we have to temporarily make the buffer writeable if we wish to
stamp our new addition into the buffer.

Since this is a change complete unto itself, we save the buffer
afterwards.